requirements.txt
beautifulsoup4>=4.12.0
wikipedia>=1.4.0
googlesearch-python>=1.2.0
youtube-transcript-api>=0.6.0
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0